rename controlnet_wrapper.py
