+++ src/streamdiffusion/acceleration/tensorrt/__init__.py
@@ -12,7 +12,7 @@ from .builder import EngineBuilder, create_onnx_path
 from .engine import AutoencoderKLEngine, UNet2DConditionModelEngine
 from .models import VAE, BaseModel, UNet, VAEEncoder
 from ...model_detection import detect_model, extract_unet_architecture, validate_architecture
-from .controlnet_wrapper import create_controlnet_wrapper
+from .unet_controlnet_export import create_controlnet_wrapper
 from .engine_pool import ControlNetEnginePool
 from .ipadapter_wrapper import create_ipadapter_wrapper
