+++ src/streamdiffusion/wrapper.py
@@ -988,7 +988,7 @@ class StreamDiffusionWrapper:
                     extract_unet_architecture,
                     validate_architecture
                 )
-                from streamdiffusion.acceleration.tensorrt.controlnet_wrapper import create_controlnet_wrapper
+                from streamdiffusion.acceleration.tensorrt.unet_controlnet_export import create_controlnet_wrapper
                 from streamdiffusion.acceleration.tensorrt.ipadapter_wrapper import create_ipadapter_wrapper
 
                 # Legacy TensorRT implementation (fallback)
